# Reuse StringLatin1::equals in regionMatches