@franz1981 2023-12-04
@schlosna 

Running `TEST="micro:java.lang.StringComparisons.regionMatches"` on AMD 7950x@4.5 GHz with tuned network-latency profile on and turbo-boost disabled using `numactl --localalloc -N 0` to avoid weird NUMA-like effects on heap objects.

baseline at 25f9af99be1c906fc85b8192df8fa50cced3474f:
```
Benchmark                             (size)  (utf16)  Mode  Cnt    Score   Error  Units
StringComparisons.regionMatches            6     true  avgt    5    4.380 ? 0.030  ns/op
StringComparisons.regionMatches            6    false  avgt    5    5.772 ? 0.056  ns/op
StringComparisons.regionMatches           15     true  avgt    5    4.005 ? 0.104  ns/op
StringComparisons.regionMatches           15    false  avgt    5    4.030 ? 0.055  ns/op
StringComparisons.regionMatches         1024     true  avgt    5   30.037 ? 0.089  ns/op
StringComparisons.regionMatches         1024    false  avgt    5   17.734 ? 0.092  ns/op
StringComparisons.regionMatchesRange       6     true  avgt    5    4.825 ? 0.067  ns/op
StringComparisons.regionMatchesRange       6    false  avgt    5    5.878 ? 0.056  ns/op
StringComparisons.regionMatchesRange      15     true  avgt    5    5.736 ? 0.069  ns/op
StringComparisons.regionMatchesRange      15    false  avgt    5    5.447 ? 0.028  ns/op
StringComparisons.regionMatchesRange    1024     true  avgt    5   31.169 ? 0.009  ns/op
StringComparisons.regionMatchesRange    1024    false  avgt    5   16.614 ? 0.168  ns/op
```
With 1bd619a5bd2faa8057cb85105b2c9b4997fbf2ac:
```
Benchmark                             (size)  (utf16)  Mode  Cnt    Score   Error  Units
StringComparisons.regionMatches            6     true  avgt    5    3.535 ? 0.022  ns/op
StringComparisons.regionMatches            6    false  avgt    5    3.134 ? 0.022  ns/op
StringComparisons.regionMatches           15     true  avgt    5    2.568 ? 0.022  ns/op
StringComparisons.regionMatches           15    false  avgt    5    3.415 ? 0.017  ns/op
StringComparisons.regionMatches         1024     true  avgt    5   30.052 ? 0.070  ns/op
StringComparisons.regionMatches         1024    false  avgt    5   17.024 ? 0.111  ns/op
StringComparisons.regionMatchesRange       6     true  avgt    5    4.819 ? 0.010  ns/op
StringComparisons.regionMatchesRange       6    false  avgt    5    5.888 ? 0.083  ns/op
StringComparisons.regionMatchesRange      15     true  avgt    5    5.849 ? 0.106  ns/op
StringComparisons.regionMatchesRange      15    false  avgt    5    5.466 ? 0.069  ns/op
StringComparisons.regionMatchesRange    1024     true  avgt    5   31.177 ? 0.015  ns/op
StringComparisons.regionMatchesRange    1024    false  avgt    5   16.872 ? 0.387  ns/op
```
Which translate in a ~1.8 speedup for small sized ones (which is still a fairly common use case), while bigger ones seems unchanged. 
This can be either due to caching effects becoming more relevant and/or a better behaviour of the equals intrinsic
for tail/remaining bytes, if compared to `mismatch`: I need to `perfasm` to check what's going on.