@franz1981 2023-12-04
The new commit, introducing the full positive use case (maybe relevant for the case with few characters), and adding
and additional comparison vs the `String::equals` case.

baseline at 25f9af99be1c906fc85b8192df8fa50cced3474f:
```
Benchmark                            (size)  (utf16)  Mode  Cnt   Score   Error  Units
StringComparisons.same                    6     true  avgt    5   2.402 ? 0.028  ns/op
StringComparisons.same                    6    false  avgt    5   2.056 ? 0.056  ns/op
StringComparisons.same                   15     true  avgt    5   3.733 ? 0.161  ns/op
StringComparisons.same                   15    false  avgt    5   2.807 ? 0.214  ns/op
StringComparisons.same                 1024     true  avgt    5  23.485 ? 0.150  ns/op
StringComparisons.same                 1024    false  avgt    5  15.302 ? 0.232  ns/op

StringComparisons.sameRegionMatches       6     true  avgt    5   4.410 ? 0.078  ns/op
StringComparisons.sameRegionMatches       6    false  avgt    5   5.414 ? 0.028  ns/op
StringComparisons.sameRegionMatches      15     true  avgt    5   5.770 ? 0.021  ns/op
StringComparisons.sameRegionMatches      15    false  avgt    5   5.771 ? 0.035  ns/op
StringComparisons.sameRegionMatches    1024     true  avgt    5  30.964 ? 0.023  ns/op
StringComparisons.sameRegionMatches    1024    false  avgt    5  16.807 ? 0.181  ns/op
```
with 1bd619a5bd2faa8057cb85105b2c9b4997fbf2ac:
```
Benchmark                            (size)  (utf16)  Mode  Cnt   Score   Error  Units
StringComparisons.sameRegionMatches       6     true  avgt    5   3.442 ? 0.016  ns/op
StringComparisons.sameRegionMatches       6    false  avgt    5   3.117 ? 0.002  ns/op
StringComparisons.sameRegionMatches      15     true  avgt    5   4.759 ? 0.075  ns/op
StringComparisons.sameRegionMatches      15    false  avgt    5   3.813 ? 0.026  ns/op
StringComparisons.sameRegionMatches    1024     true  avgt    5  28.308 ? 0.058  ns/op
StringComparisons.sameRegionMatches    1024    false  avgt    5  16.774 ? 0.220  ns/op
```
As confirmed by the previous results, the better value of this PR is with small-sized strings, but it is yet to be verified if:
- due to a better tail-handing of the `equals` intrinsics
- larger strings are limited by the amount of cache activity

Running `perfnorm` against 25f9af99be1c906fc85b8192df8fa50cced3474f with `same` vs `sameRegionMatches` reveal that 
both have high IPC with an high number of branche AND L1 cache-loads (with nearly no misses), which means that the amount of branches and cache-loads dominate the cost of computation despite the number of instructions on `sameRegionMatches` is higher.
In short, with bigger sized-strings, the difference between the 2 instrincs just fade away.

Now forcing some tail execution on mid-long sized ones to verify if there's some better handling of tail bytes from by `same`.