@franz1981 2023-12-04
@cl4es I can undraft this, but I have no powers to create a JDK issue myself :"/