@cl4es 2023-12-04
General comments:
- Does this suggest an optimization opportunity to `ArraysSupport::mismatch`, or is the relative overhead due the inherent costs of finding the exact index in the mismatch? 
- This only optimize the case where we're comparing the entirety of both `String`s. How common is this in practice? It would be good to check a mixed benchmark where some inputs will use this fast path and others won't to assert that this doesn't have an unexpected negative impact from adding a few tests.