@cl4es 2023-12-04
Filed: https://bugs.openjdk.org/browse/JDK-8321283